x86/checkasm.asm: deduplicate / simplify some code
We can use [stack_param] to automatically account for the shadow space
when needed, and we don't need to store a copy in r5 because we can
reconstruct it at any time. This also lets us load the comparison variable
directly in r1 to save another mov.

Finally, the null byte can be written in the platform independent section.