riscv/checkasm_64.S: improve stack canary check
The current logic was a bit strange. It calculated the address of the stack
canary, but then:

1) kept the stack canary itself unchanged
2) wrote the inverse of its value to the stack canary reference location

When comparing, it then compared that location against the inverse stored
in the stack canary reference. It would be better to actually *clobber*
the stack canary itself, so that any write to (sp) will trigger the stack
canary failure.

After this change, we handle this by instead writing the (inverse of the)
address of the stack canary itself to the stack canary location, ensuring
it will detect almost any change/use (e.g. writing a 0). This also means
we don't need to store a second value anymore to compare against.

(Note: We also don't need to allocate extra space on the ARG_STACK for the
stack canary, because the only thing we could clobber is our own extra
arguments from the argument counting boilerplate)