checkasm: dramatically simplify CPU flag management
Instead of *requiring* two separate callbacks just for managing CPU
flags, we go for the much simpler design of having a single field for the
detected flags (set by the user), and providing a single callback for the
user to get the currently masked set of CPU flags.

Users can already forward this to their subsystems as needed, or we could
continue offering a convenience callback to do it for them.

Also rewrite the internal CPU flag tracking logic to be simpler and require
less traversal of the CPU flag list by storing a pointer to the CPU flag
info structure directly.