x86/checkasm.asm: don't return after first error
A function can have multiple mistakes, such as clobbering a register _and_
leaving the x87/YMM in a bad state. In such a case, we should check for all
such errors, and also clean up after the asm function.

It's worth pointing out that our code does still allow for the possibility
of checkasm_fail_func() itself being called with bad x86/YMM state, but
this should hopefully be a non-issue in practice as this function will
basically just call fprintf with integer arguments.