perf: measure perf number scale
So we can also convert these to ns for perhaps more intuitive results,
as well as for internal uses (e.g. targeting a specific time budget per
function).

It's a bit tricky to get a reliable calibration for the perf numbers on
all platforms, so I went with a very general algorithm that assumes as little
as possible about the precision of gettime(). We first determine the number
of noops needed to hit a ~20 µs target time, and then measure the perf
counter for that duration. At the same time, we measure a wallclock reference,
with a separate loop to avoid the performance overhead of the perf code
itself impacting the measurement.

This allows us to (roughly) calibrate the cycle/tick numbers against a
wallclock reference.