checkasm: try and gather enough samples to perform statistics on
It's a bit complicated in that we're sampling what is itself a mean of
samples, so we're effectively computing a mean of means (of means). But
empirically, 8 seems to be a good hard minimum and 10-20 is a more ideal
figure.

(Realistically, we well usually gather more anyways, since for small
functions we start off slow)

This does slightly extend the runtime of checkasm on e.g. dav1d, but not by
a huge amount.