include/test.h: actually use func_type typedef
There is no real reason to have func_ref/func_new be plain void pointers if
we already have a typedef. Saves a few unnecessary casts in other places.

As an aside, there's no reason to redefine func_new as tfunc in the bench
code - func_new is already a local variable.

Avoids an "unused typedef" warning as well.