checkasm: don't use symbol alias for fail_internal
It's really annoying that we have to share this same symbol between
external and internal code, but it is what it is. Insteal, just
define the function twice.

Alternatively, we could switch to external call patterns inside all
asm, but that change is more likely to break than this one.