x86: make emms check more robust
This check relied on the fact that x86 CPUs seem to set the high bits of
the written control words to -1. However, this behavior is not documented
and also e.g. does not hold when running inside QEMU.

Instead, just load the 16-bits and compare them against the expected value.