stats: switch to log-normal distribution
This is a much better fit for timing data, which is strictly positive and
typically has a fat tail with multiplicative noise.

In fact, using a log-normal distribution alone makes the raw mean so much
more reliable that we don't even need to trim outliers to get a clean signal,
thus vastly simplifying our overall measurement process.

Fixes: #2