checkasm/test: always define fallback perf implementation
Since we have a gettime_nsec() anyway, we might as well use it to provide
a baseline for all platforms.

This implementation pre-emptively defines a struct to hold the actual fallback
implementation, since I plan on moving the Linux perf and macOS kperf
implementations to this mechanism as well.

Allows us to make the timing code unconditional.