include/perf: define CHECKASM_PERF_ASM_USABLE
This is set to 1 if we can guarantee that CHECKASM_PERF_ASM is always
usable (i.e. doesn't require a test).

I decided to set it only on platforms I have access to and/or can confirm
that this is the case. If it is true for loongarch or ppc64le as well,
and you are reading this, please add it.