checkasm: don't stop sample growth prematurely
Otherwise, we risk biasing towards a certain number of iterations.

Now that we grow more slowly (~1% instead of ~12%), there is not really an
issue with just gathering 30 samples. For very slow functions, overall runtime
should actually still be even better as a consequence of the fact that we
no longer round up to 32 calls, either.