internal: add mach-based checkasm_gettime_nsec implementation
clock_gettime(CLOCK_MONOTONIC_RAW) on macOS 10.12+ is just an alias for
mach_continuous_time(). On older versions, clock_gettime() is not implemented.

This implies that we are safe to simply use mach_absolute_time() (basically
the same thing as mach_continuous_time for our purposes) on all macOS
versions.