perf: simplify CHECKASM_PERF_ macros
CHECKASM_PERF_SETUP() can be just inlined as needed, since its contents
are now basically unchanging, and we can also avoid the need to have
separate macros for the name/unit.

I also want to hide the implementation of the underling timed call behind
a helper macro (CHEKCASM_PERF_BENCH) for an upcoming change, so do that now.