perf: only use perf ASM if available
perf.start/stop() calls can be extremely slow (e.g. when they involve multiple
syscalls). In this case, it's better not to run a timer call on every single
function invocation, but rather to group together multiple invocations in a
naive loop.

For typical bench durations, this should still be short enough to not trigger
any scheduling headaches.

This commit both enables the runtime testing of ASM availability, and also
vastly improves the performance of checkasm with non-ASM timer sources.