perf: switch to naive loop when not using perf asm
perf.start/stop() calls can be extremely slow (e.g. when they involve multiple
syscalls). In this case, it's better not to run a timer call on every single
function invocation, but rather to group together multiple invocations in a
loop.

For typical bench durations, this should still be short enough to not trigger
any scheduling headaches.

Vastly improves the usability of checkasm with non-ASM timer sources.