meson: Enable more warning flags
These are quite valuable warning flags, pointing out potential
issues; most of these flags are enabled in dav1d.

-Wundef allows catching build configuration issues, when referring
to defines that aren't set.

-Wshorted-64-to-32 points out some cases of unintentionally truncating
64 bit values to 32 bit that can be real issues. MSVC warns about this
by deault (in our current warning level configuration) but we have this
warning disabled, as that warning (C4244) also warns about a number of
other things that we might not care about.