checkasm: keep track of test and report name for each function
Since we don't know the report name until *after* checkasm_report() is called,
we need to retroactively keep track of which functions correspond to the same
group of reports. This is done by linking together functions that were
initially allocated as part of the same group.

I decided to `strdup`the report name onto every single function, to save
needing an extra `int` to track which function owns which string.

If C++ browsers can handle millions of allocations per second, C can handle
a few thousand.