checkasm: subtract nop cycles after averaging
Since we use a geometric mean, it's better for us to subtract the nop cycles
after averaging, since that avoids the possibility of the overall product
exploding towards 0 when the measurement series includes a single noop term.

This also helps with an upcoming change, where I want to measure the noop
over multiple iterations throughout the lifetime of the program, to get a
more stable measurement, but that conflicts with using it at the time of
the measurement itself.