checkasm: also print nop time as part of function listing
This partially reverts a previous change, where the nop time printing was moved
to the header. The main motivation here is that I want to aggregate the nop
cycle count multiple times over the course of the benchmark. By reporting it
a second time we inform the user about the true nop overhead.