checkasm: add HTML output option
The basic layout and output template was unceremoniously stolen from
haskell-criterion, under MIT license, and heavily modified for our
purposes.

To get the data into the library, we use C23 #embed, which means this
feature only works if __has_embed is defined. That does mean I have to
enable -Wno-c23-extensions, or we get a warning every time.