perf: simplify perf scale calculation
Instead of measuring the cycles/iter and time/iter independently,
measure both on the same loop and run the regression directly on the
Y vs X axis where the Y axis is nsec and the X axis is cycles.

Gives the same answer in the end but in a less round-about way.