platform/x86_64: fix off-by one in arg counting
It's possible for a function to have 0 regular arguments, e.g. in the case of
all arguments being float (and hence passed via XMM registers). To solve it,
add a 0 entry after the last argument. To avoid changing the total number of
arguments, we can just remove the 16 value instead, since the existing code
assumes a maximum of 15 arguments anyway.