tests/generic: don't use 0xFF in memory corruption tests
Our self-check asserts that we correctly detect test data mismatches. This
requires that we actually hit the failure path, which means any previously
harmless false negative errors would now fail the CI check (turning into a
false positive).

In the past, such false negatives were pretty unlikely to be witnessed
in practice, but still possible. With the new path, setting a buffer to
"all 1s" is a specific pattern that checkasm is likely to test, meaning we
are much more likely to get a false negative here if we recreate the same
test pattern accidentally.

Set the buffer to some random byte instead, to avoid this issue, and also
write multiple bytes to make it more robust.