Avoid directly referencing data symbols across the libcheckasm API
When directly accessing a data symbol (as opposed to calling a
function) in a different DLL, MSVC requires the symbol declaration
to have a dllimport attribute.

(In mingw environments, one can do without dllimport attributes,
thanks to a mechanism called runtime pseudo relocations.)

Note - the dllimport attributes need to be in place when the
_caller_ of the API parses the headers. This has the implication
that one set of headers only works specifically for one library
configuration (shared or static), not both.

(Usually this requires either defaulting to either configuration,
installing a config header along with the regular headers, or
requiring the user to set a define when including the headers, to
choose the linking mode.)

To avoid the burden of having to juggle the dllimport attribute,
avoid directly accessing data symbols in the public library
interface - instead use accessor functions for such symbols.
accessed over a DLL boundary, in the case of MSVC builds.