meson: Don't use dllexport for DLLs on mingw on x86_32
When exporting a symbol using dllexport attributes, the compiler
places an "-export:<name>" directive into the object file, which
the linker then reads and uses for deciding which symbols to
export.

Symbols on x86_32 on Windows have an underscore prefix,
so a C level function "func" has the symbol name "_func".

However - there is an internal inconsistency between how mingw and
MSVC style toolchains handle this; MSVC style object files include
the underscore prefix in the export directive in the object file,
e.g. "-export:_func", while mingw style object files don't
expect that leading underscore there.

When building assembly files with nasm, it doesn't know which
of these conventions the surrounding build environment uses, and
just uses the style of MSVC toolchains.

Hence, it's impossible to use nasm's dllexport feature in a
mingw environment on x86_32.

Therefore, just skip setting the defines for doing dllexport
on mingw x86_32. This allows the toolchains to fall back to the
default behaviour of exporting all symbols, which should work fine
(albeit suboptimal).

With GNU toolchains, there is another small complication regarding
this: When exporting all symbols without explicit dllexport, the
linker needs to decide whether a symbol is a function or data.
GNU ld does this based on the symbol type. With GNU assemblers,
one can set this with ".def func; .scl 2; .type 32; .endef" to mark
the symbol as a function. Nasm doesn't seem to have a way of
doing the same.

Therefore, GNU ld thinks all the symbols exported from nasm object
files are data symbols, not functions. When these are referenced
without dllimport, it causes the linker to use runtime pseudo
relocations to link them. In 64 bit mode, this can be fatal, as
the function call sites use a 32 bit relative offset, which can
be out of range when DLLs are loaded too far away in the 64 bit
address space. This should not be an issue in 32 bit mode though.

Therefore - in short, this gotcha is benign in x86_32 mode, which
is the only case where we skip the explicit dllexports.