arm: Fix the checked_call wrappers for handling only float arguments
On ARM with hardfloat ABIs, and on aarch64, float arguments are
passed in different registers than regular arguments.

This means that a function that only takes a float argument and
nothing more, would end up using no GPRs for the arguments.

The existing invocations of the checked_call functions were
designed assuming that there would be at least one GPR parameter,
with respect to the tail of numbers of parameters passed on the stack.
Add one number to each of these invocations, to cover the case when
there are no GPR parameters at all.

This fixes crashes in the float argument tests on arm and aarch64.