checkasm: handle signals internally; re-run entire test on signal
The motivating reason for this change is to allow calling declare_func() and
check_func() out of order, however it has several other advantages:

- The longjmp implementation can become private and depend on config.h in a
  straightforward way; also requiring less meson logic

- It categorically avoids any possible clobbering of local state by longjmp,
  even if e.g. the tested function modifies a variable not marked as volatile

- We can easily make our signal handling thread-safe in the future

The only downside is that a failing test will go through its internal setup
logic additional times; but other than that, already-failed tests will be
silently skipped, essentially leading the called test() straight back to the
next function to check; same as before.

We also need to be a bit careful in how we continue to associate report()
calls with checked functions after the signal handler has been invoked, which
does require that we keep track of the number of reports in the current test
group.