meson: Detect support for assembling SVE instructions
We don't need the .arch directive for enabling SVE (it's only
required for enabling certain instruction sets like dotprod
and i8mm, only with binutils, and only with binutils 2.40 and
older).

Move '@EXTRA_ARGS@' to the generator object only, we don't want
that included when running the command manually during configure.

The detection of SVE support in the toolchain is somewhat unwieldy
(when covering support for gas-preprocessor for MS armasm as well),
but allows building those codepaths when suitable.

This is also quite optional; we only need it for printing an info
notice at startup about the vector length. (We don't need the
info for picking a specific version of the checked_call wrapper.)