signal: handle SIGINT/SIGTERM cleanly
When encountered, we should print a summary of the work we did so far. To
handle the signal cleanly, we just raise an internal flag and then wait for
the test framework to finish the current function.

We resolve the interrupt after each report() call or before starting a new
test.