aarch64: Add runtime checking of support for SVE
This slightly deviates from the previous design direction; it was
probably intended to have checkasm_sve_length() check for support
for SVE internally, just like checkasm_init_riscv() does above.

However, I prefer having the separation more clearly like this
visible on the C level, with smaller scoped assembly functions
with less logic in them.

This is mostly an optional feature; we only need it for printing
an info notice at startup about the vector length. (We don't need
the info for picking a specific version of the checked_call wrapper.)
But it's neater to have this integrated into checkasm itself, rather
than requiring the user applications to check and print it themselves
(e.g. in a callback).