tests: Implement aarch64 specific tests
Test that the wrapper function catches clobberings for all relevant
registers (and that all other registers can be clobbered without
error), both for GPRs and float/vector registers.

Also check that we correctly allow clobbering the upper half of
registers d8-d15.

Check that the invocation of functions actually does force injecting
nonzero bits in the upper half of registers for 32 bit parameters.

Check that we catch illegal instructions.

Check that we can clobber the part of the stack where arguments are
passed (assembly functions rarely do this, but compiler generated
reference functions can do it at times - see
55cf967bdf6f94d963080f57c34271ef0623f68f in dav1d for where this
was introduced). And check that if we clobber the stack outside
of the region of stack arguments, the call wrapper does detect it.