tests: Implement tests for 32 bit arm
This reuses a couple helpers from libcheckasm - make them visible
in a shared library build.

This checks mostly the same as the 64 bit tests, except for details
that aren't relevant in 32 bit mode.

Additionally, check that we do detect clobbering the fpscr register,
which this checked_call wrapper does check for.