function: add new helper file to contain function tree implementation
This code is logically distinct enough from checkasm.c and contains a lot
of internal helpers that don't really matter from the point of view of the
rest of the checkasm code, so I think it makes sense to separate it out.

In theory, we can go a step further and also allow for separate function trees
which we can then merge with the main function tree, in case we want to be
able to run them in parallel.

I also opted to hide the pointer-to-root (i.e. the tree itself) behind a
separate type just to reduce the prevalence of a double pointer.