perf: Try using an ARM11 (ARMv6) specific timer
Try using this timer through the checkasm_perf_init_arm
case, just like we try using cntvct_el0 on aarch64.

This uses the same pattern as the setup/testing for the main
assembly timer in src/perf.c; we check if we can access the
register, we try to optionally start the timer (on all cores),
and check that it progresses.

This provides a working cycle timer on Raspberry Pi 1.