perf: Try to enable the ARM timers, if possible
The cycle counter timers on ARM can be stopped/started. In
particular, using Linux perf leaves the timers stopped afterwards.

However, depending on configuration (on the kernel side), it
can also be in a state where the timers can't be started/stopped
from user space at all. (On Windows, the timer is always
available, and can't be started/stopped.)

Allow the perf headers to define a function that can try to
enable the timers. Run this with a signal handler, in order to
gracefully skip it if we aren't allowed to do this.

If running the enable routine succeeded, run the enable routine
on all cpu cores. The implementation of checkasm_run_on_all_cores
doesn't have a Windows codepath yet, as we don't really need to do
this on Windows at the moment.

This is based on control code by Janne Grunau from
https://code.videolan.org/janne/arm64-cycle-cnt.