function: simplify get_func() helper
Instead of relying on the implicit magic value of f->func to see if the
node was newly allocated or not, just track this directly in the return
value. A bit simplier and requires fewer pointer indirections on the way up,
as well as only updating the root node when actually necessary.

Arguably makes the code simpler to understand, and may help with making this
thread safe in the future.