perf: skip measuring perf scale if the timer unit is "nsec"
While it could be nice to verify this scaling in theory, in practice this
is currently only ever the case if the perf backend is either provided *by*
gettime_nsec() (in which case measuring the scale of one using the other
is pointless), or if we are using the 1 GHz timer register on ARM - but we
know the frequency of this one with certainty.

The only other time this is really useful is when validating the perf scale
measurement code itself; but it's easy enough to just temporarily disable
this branch in that case.