include/test.h: use static variables for func_ref/func_new()
This is required to allow setting these function pointers (via check_func)
before calling declare_func().

We have to do a bit of pointer type juggling to be able to use these
static void * variables as lvalues; so that users can continue overriding
func_ref/func_new as before.

The major downside of this approach is that we lose the ability to catch
type mismatches on the function pointer passed to `check_func()`, e.g.:

  void (*myfunc)(int);

  declare_func(void, int, int);
  if (check_func(myfunc, ...))
      ...

While we could reintroduce this in the specific case that `declare_func` is
called before `check_func`, due to fundamental limitations of the C macro
preprocessor it will necessarily require shadowing some sort of global
typedef, which carries with it the risk of raising shadowing warnings.

Overall, losing this warning is not a big deal, and may even sometimes be
intentional (e.g. when implicitly aliasing pointers or testing variadic
functions).