checkasm: handle signals internally; re-run entire test on signal
The motivating reason for this change is to allow calling declare_func() and
check_func() out of order, however it has several other advantages:

- The longjmp implementation can become private and depend on config.h in a
  straightforward way; also requiring less meson logic

- It categorically avoids any possible clobbering of local state by longjmp,
  even if e.g. the tested function modifies a variable not marked as volatile

- We can easily make our signal handling thread-safe in the future

The only downside is that a failing test will go through its internal setup
logic additional times; but other than that, already-failed tests will be
silently skipped, essentially leading the called test() straight back to the
next function to check; same as before.

We also need to be a bit careful in restoring the intra-report state correctly,
but there is a pretty elegant solution here: when a function crashes, we
mark it as a special state `CRASHED`, and then re-associate any prior
failures with the current function once we see that state again.

It's worth pointing out that there is an implicit risk that the resumed
test does not exactly match the same order of check_func() calls as the
interrepted run; but this risk should be greatly mitigated by the fact that
we re-seed the PRNG on every `check_func()` or `report()` call - and arguably
such tests are broken in the first place. Still, we can at least detect a
subset of such cases and error out.

Closes: #5