arm: Fix the clobber_stack asm functions for Apple platforms
On Apple platforms, int parameters on the stack are packed tightly,
not laid out in a 64 bit slot for each parameter like on Linux and
Windows.

This caused the clobber_stack_args_aarch64 function to actually
clobber part of the caller's stack. On Apple platforms, we don't
have a working checked_call wrapper, so this directly would
affect the calling, compiler generated function's stack.

Also, for non-Apple platforms, only overwrite the 32 bit part
of the actual parameters on the stack, not the whole 64 bit
parameter slot.

Fixes: #19