arm: Implement a checked_call wrapper for Apple platforms
We couldn't use the same checked_call function on Apple platforms
as on other platforms, as 32 bit arguments are packed tightly on
the stack, so when passing 8 dummy arguments in registers, we don't
know whether we should read 4 or 8 bytes per argument from the stack
into each register, when shifting arguments back into registers.

Initially, when Apple on arm64 only was iOS and similar platforms,
it wasn't a big deal that checkasm didn't have all the check
ergonomics as elsewhere, but these days, we want to provide as
many checks as possible for developers on macOS.

Instead implement the checked_call wrapper slightly differently.
This approach relies on passing the __VA_ARGS__ for the real
parameters first, so they don't need to be shifted, while passing
all other parameters (including func_new) after that.

Surprisingly, the ARG_STACK define needs one extra set of
parentheses to get the right evaluation order, when assembled
for Apple targets.