riscv: best effort to clobber argument registers
To support forward-edge CFI, the prototype of the function (pointer)
used to check calls must be the exact same as that of the actual
function being checked. As such, it will no longer be possible to rely
on the checked_call long argument list to clobber unused argument
registers.

This adds a best effort ABI-compliant method to clobber registers as
late as possible, i.e. right before the checked all. In practice, it
works because the compiler has no reasons to set unused argument
registers at that point.