cpu: Print the type of CPU cores on ARM on Linux
Locate the "CPU implementer" and "CPU part" fields from /proc/cpuinfo.

These fields are opaque integers; use tables for looking up the
name of the implementer, and the implementer specific interpretation
of the part number.

Print one instance of each individual kind of core present. As an
example, on a RK3399, it prints this:

    - CPU: ARM Cortex-A53, ARM Cortex-A72

The list of cores is gathered from cores I have access to.