docs: add doxygen build process
This is a bit ugly on account of the fact that doxygen *insists* on ingesting
the list of input files via the Doxyfile, rather than the command line (which
would be sensible).

So we have to generate a properly escaped list of source file paths, which
requires preserving the full file paths from subdir('include') before
converting them to file() objects.