Remove nb_* fields from CheckasmConfig
This is, of course, absolutely not backwards compatible, but we decided in
the discussion to ignore that fact due to the fact that literally no downstream
project depends on checkasm v1.0.0 (released just a few days ago) yet.

We'll just retroactively call this the always-intended behavior and the
change a bugfix, and include it as part of v1.0.1.

Closes: #23