include/test.h: avoid strict aliasing violation
Unfortunately, this is fundamentally impossible to avoid without breaking
the ability to use `checkasm_func_ref/new` as an l-value. Since we have
the generic `checkasm_call()` and `checkasm_call_checked()`, however, there
is very little reason for users to override these function pointers anymore.

This is being done currently by a single test in FFmpeg, and that one would
be easy to update.

Fixes: #26