include/test.h: expose checkasm_key_ref/new
And use these instead of the (previously hidden) checkasm_ptr_ref/new.
Useful for when the user needs to cast them to a different function type,
e.g. for use in (checked) wrappers.

Since this makes `checkasm_check_key` now almost wholly redundant, demote
that to an internal-only function. This does not break any existing ABI.

The change to include static variables in Doxyfile.in is to ensure that
these variables are actually visible.

Fixes: #24