Abort on internal failures, rename to checkasm_fail_abort()
This function is typically called by checked wrappers as a result of ABI
violations, register corruption, stack corruption etc.

When this happens, there is really very little guarantee that the program
state is still valid; continuing execution is likely to lead to further
problems. It's much more principled to just longjmp back to the signal
handler in this case, and treat it like a normal exception.

Also fixes a related issue where the RISC-V harness was not preserving
return values after failing a function.

Fixes: #25